+++ src/entrypoint.py
@@ -46,13 +46,18 @@ def run_coveralls(repo_token, parallel=False, flag_name=False, base_path=False):
     # for some reasons the "service_name" can be one or the other
     # (depending on where it's ran from?)
     service_names = ("github", "github-actions")
+    # sets `service_job_id` key so it exists, refs:
+    # https://github.com/coveralls-clients/coveralls-python/pull/241/files#r532248047
+    service_job_id = None
     result = None
     if base_path and os.path.exists(base_path):
         os.chdir(base_path)
     for service_name in service_names:
         log.info(f"Trying submitting coverage with service_name: {service_name}...")
         with patch_os_environ(repo_token, parallel, flag_name):
-            coveralls = Coveralls(service_name=service_name)
+            coveralls = Coveralls(
+                service_name=service_name, service_job_id=service_job_id
+            )
             try:
                 result = coveralls.wear()
                 break